Added Toggle function in TaskList
